Revert "[RN-781]: ios/android quick links"
