+++ android/settings.gradle
@@ -1,6 +1,4 @@
 rootProject.name = 'BitPay'
-include ':react-native-quick-actions'
-project(':react-native-quick-actions').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-quick-actions/android')
 include ':react-native-appsflyer'
 project(':react-native-appsflyer').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-appsflyer/android')
 include ':react-native-haptic-feedback'
