+++ ios/BitPayApp/AppDelegate.m
@@ -14,7 +14,6 @@
 #import <FlipperKitNetworkPlugin/FlipperKitNetworkPlugin.h>
 #import <SKIOSNetworkPlugin/SKIOSNetworkAdapter.h>
 #import <FlipperKitReactPlugin/FlipperKitReactPlugin.h>
-#import "RNQuickActionManager.h"
 
 #if RCT_DEV
 #import <React/RCTDevLoadingView.h>
@@ -94,10 +93,6 @@ static void InitializeFlipper(UIApplication *application) {
   } else {
     completionHandler(UNNotificationPresentationOptionAlert);
   }
-}
-
-- (void)application:(UIApplication *)application performActionForShortcutItem:(UIApplicationShortcutItem *)shortcutItem completionHandler:(void (^)(BOOL succeeded)) completionHandler {
-  [RNQuickActionManager onQuickActionPress:shortcutItem completionHandler:completionHandler];
 }
 
 - (NSURL *)sourceURLForBridge:(RCTBridge *)bridge
