+++ package.json
@@ -113,7 +113,6 @@
     "react-native-progress": "5.0.0",
     "react-native-prompt-android": "1.1.0",
     "react-native-qrcode-svg": "6.1.2",
-    "react-native-quick-actions": "0.3.13",
     "react-native-randombytes": "3.6.1",
     "react-native-rate": "1.2.9",
     "react-native-reanimated": "2.5.0",
