+++ src/components/tabs/Tabs.tsx
@@ -1,13 +1,9 @@
-import React, {useEffect, useMemo, useState} from 'react';
-import {DeviceEventEmitter, View} from 'react-native';
+import React, {useMemo, useState} from 'react';
+import {View} from 'react-native';
 import styled from 'styled-components/native';
 import {ScreenGutter} from '../styled/Containers';
 import TabButton from './TabButton';
 import {useAppSelector} from '../../utils/hooks/useAppSelector';
-import QuickActions, {ShortcutItem} from 'react-native-quick-actions';
-import {useAppDispatch} from '../../utils/hooks';
-import {shortcutListener} from '../../store/app/app.effects';
-import {useNavigation} from '@react-navigation/native';
 
 interface TabsProps {
   tabs: () => {
@@ -27,8 +23,6 @@ const Tabs: React.VFC<TabsProps> = props => {
   const {tabs} = props;
   const [activeTabIdx, setActiveIdx] = useState(0);
   const defaultLanguage = useAppSelector(({APP}) => APP.defaultLanguage);
-  const dispatch = useAppDispatch();
-  const navigation = useNavigation();
 
   const memoizedTabs = useMemo(() => {
     const tabData = tabs();
@@ -52,20 +46,6 @@ const Tabs: React.VFC<TabsProps> = props => {
       {d.title}
     </TabButton>
   ));
-
-  // QUICK LINKS
-  useEffect(() => {
-    QuickActions.popInitialAction()
-      .then(item => dispatch(shortcutListener(item, navigation)))
-      .catch(console.error);
-    const subscription = DeviceEventEmitter.addListener(
-      'quickActionShortcut',
-      (item: ShortcutItem) => {
-        dispatch(shortcutListener(item, navigation));
-      },
-    );
-    return () => subscription.remove();
-  }, []);
 
   return (
     <View>
